tests/requirements-test.txt
# Test dependencies for RLS/RBAC test suite
# Install: pip install -r tests/requirements-test.txt

pytest>=7.4.0,<9.0.0
psycopg2-binary>=2.9.7,<3.0.0
python-dotenv>=1.0.0,<2.0.0
